requirements.txt
numpy>=1.19.0
opencv-python>=4.5.0
pillow>=8.0.0
torch>=1.7.0
torchvision>=0.8.0
ultralytics>=8.0.0
pandas>=1.3.0
openpyxl>=3.0.0
face-recognition>=1.3.0